Partially revert 0772ea8c32198901505cde988d4374787f77c69d
As it turns out, removing the ids for snapshoting at the entity and vertex level break the undo-manager.
So I had to reinstate the ugly thing to do it at the level level.

Also, cloneEntity now uses the entity processor, which fixes the bug that the previous commit was attempting to solve.