Fix mobile bug where the virtual keybord makes the mode bar disapear
Basically, when the virtual keyboard appears, the input is scrolled into view, and the mode bar goes of screen. When the keyboard is toggled off, it doesn't scroll back up in view. This is fixed by using flexbox.

While doing that, I noticed that a bigger canvas than necessary was being drawn on, parts of which were off-screen. I fixed that by using position: absolute in combination with flexbox, as described above.
I expect that the editor will be a tiny bit faster because it will have to render a bit less stuff.

More context: https://discord.com/channels/526188679006388254/601609930486841365/885931073672974386